Extract drag drop relocation
