+++ src/gui_app/folder_browser.rs
@@ -362,6 +362,7 @@ mod folder_entry;
 use folder_entry::FolderEntry;
 
 mod drag_drop;
+mod drag_drop_relocation;
 
 mod delete_workflow;
